Split optree optimizer and finalizer from op.c into new peep.c
 * Create a new `peep.c` file

 * Move the functions related to optree optimisation and finalisation
   out of `op.c` into this new file

 * Several previously-static functions now have to be non-static and
   declared as internal API in order to be shared between these two
   files.
